Add Settings page with volume controls
Add two audio buses, Music and Sfx, and assign all audio nodes in the
project to the appropriate bus.

Add a Settings page with one slider per bus. Persist the values to a
config file, and restore them when the Settings scene is instantiated
(which happens when the game is launched).

Co-authored-by: Will Thompson <wjt@endlessos.org>