Add inert Player to first and last levels
Previously, a Player scene would be instantiated, without its script, at
a fixed location in the first and last levels of each world, in code.
This is inflexible in the case where the world designer wants to place
this player somewhere else in the level.

Instead, instantiate the Player scene into each first and last level,
again with its script detached. Remove the special case from the
Level.gd scene.

Whereas previously the Player added in code would be set to be looking
left or right with 50% probability, with this change they will always be
looking at whichever direction they are set to in the scene. I think this
is an acceptable loss.

Helps https://github.com/endlessm/everlasting-candy/issues/31