TileSet: Remove Player/Goober sprite source & conversion code
Now that all levels have been converted to use the
TileSetSceneCollectionSource for these actors, we can remove the old
static sources from the TileSet, as well as the code that converts them
to live scenes at runtime.

Notice that the source IDs for the subsequent sources do not change,
which means the TileMapLayer nodes do not need to be updated.