World: Don't advance to level 1 twice
Consider the following steps:

1. Start world
2. Press jump on title screen to advance from level 0 to level 1
3. On level 1, die
4. During the 1.5-second timeout where “YOU LOSE” is shown, before the
   title is re-shown, press jump

The expected result (IMO) is that nothing happens. Once the title screen
is shown again, pressing jump should start level 1 afresh.

But what would previously happen is that the code path for “jump pressed
while on title screen” would be followed, starting level 1. But then
once the 1.5-second timeout is complete, level 1 would be re-started. This
is because `level` is adjusted before the 1.5-second timeout begins.

To fix this, in World's input handler, check the level_type of the
currently-displayed Level, rather than the `level` variable, when
deciding whether to act on the jump action.