Cgm-Dali: Stop timers when player wins or dies
Previously, the timer would continue running after the player has won
the level or died.

In the case where the player has died, this was harmless: the timer
would keep running to zero, but noting bad would happen.

In the case where the player has won the level, this was problematic: if
the player won the level with less than 1.5 seconds remaining, then
while the “level complete” message was showing, the player would die
when the timer expired and have to attempt the same level again. This
was frustrating and, at least for me, made the final level impossible.

In the timer, monitor for the case when the player has won or died; when
this happens, stop and hide the timer.

This could be made cleaner if Level.gd knew about the concept of the
level having a time limit: if that were the case, the level could just
stop the timer, rather than having to repeat the conditions here. But it
is a relatively small amount of code to do it here.