Cgm-Dali: Add player and platform to title and win screens
Prior to commit a862c46f7cbd8f34d19273129d97952e19857ae6, the Title.png
asset contained the central platform on which the player stands; and
prior to commit f6b196e8d3bb869608dab2aac2efb239417921a6 an inert Player
instance would be instantiated in code in the title and win levels. Now
neither is the case.

Update this world for those changes. I chose to make the central
platform out of breakable bricks, which are the unique feature of this
world.