Cgm-Dali: Replace icebreak sound effect
The previous sound is based on
<https://uppbeat.io/sfx/video-game-magic-spell-ice-shatter/1781/4093>.

Unfortunately, based on my reading of the user agreement
<https://uppbeat.io/user-agreement> we cannot use this:

> If you are a Free User, you must be an individual or a freelancer and
> you may use the Content only in material published by yourself. If you
> are a Paid Subscriber to Uppbeat Premium, you must be an individual or
> an organisation that has 5 employees or less, and you may only use the
> Content in material published by yourself or your organisation. If you
> are a Paid Subscriber to Uppbeat Business, you may be an organisation
> of any scale and may use the Content in material that will be
> published by a third party client.

Endless is not an individual or freelancer, and has more than 5
employees.

Replace it with a sound effect licensed under CC-0, which I trimmed and
lightly boosted the gain of in Ardour. Add a credit to the README, even
though it is not required by the license.