Add Settings global to handle volumes
Previously, the volume settings were restored whenever SettingsPage.tscn
was instantiated. When running the whole game, this is on startup, and
then redundantly whenever the player returns to the title. However, when
running an individual level or world from the editor, the volume
settings were not restored, making the developer sad if they are tired
of the background music.

Add a new global script, Settings, which restores the channel volumes on
_ready(), and exposes API to fetch & adjust the volume for a channel.
Adapt SettingsPage.gd accordingly. As a result, the volumes are restored
when the application starts – regardless of whether that's the whole
game or any specific scene – and only then.