CandySpawner: Make candy_texture property an array
Previously, all the candy falling in the background would have to be one
specific texture.

With this change, it is possible to specify more than one texture. They
will be used in a different order each time, cycling through the list to
avoid repetition until it is necessary.

Because CandySpawner maintains a cache of candy sprites, this doesn't
mean that every candy that falls will be random. On early levels where
only one candy is visible at a time, it will be the same one on repeat;
once the player gets far enough for two to be visible at once, a second
one will be added to the mix; and so on. But on each playthrough, the
exact textures that get used will be different.