World: Move _list_scenes() to new helper class
We will, in due course, need a new variant of this function that finds
image files.

In preparation, move it to a new ResourceFinder class. This function,
and its future sibling, are essentially polyfills for a new function
that will ship with Godot 4.4, so putting them in one place will make
it easier to clean up later.