ResourceFinder: Add function to find world-specific textures
To help build interest in getting to know Godot in an introductory
workshop, we want to make it possible to override a world's image assets
textures by just placing image files at particular paths in the world's
directory, without needing to adjust anything in the editor itself.

Add a helper function which looks up images in a named subdirectory of
an Image directory adjacent to the current scene.