CandySpawner: Find textures in world directory
Previously, to override the texture(s) used by a world's CandySpawner
instance, you would use the inspector to set the candy_textures property
on the instance. The default value of the property was a one-element
list of the default Candy.png image.

To help build interest in getting to know Godot in an introductory
workshop, we want to make it possible to override a world's candy
textures by just placing image files at a particular path in the world's
directory, without needing to adjust anything in the editor itself.

Use ResourceFinder.load_world_assets() during CandySpawner's
initialisation if the candy_textures property is empty. If no textures
are found in this directory, fall back in code to the default texture.
Change the default value of the property to the empty list.