Goober, Player, Explosion: Load textures from world if present
Previously, to override the texture(s) used by a world's Goobers or
Player, you would need to:

1. Copy the Goober/Player scene into the world (perhaps as an inherited
   scene)
2. Adjust its Sprite2D's texture
3. Copy the TileSet into the world
4. Replace the Goober/Player scene in the TileSet with the custom
   instance

Similarly for the Explosion scene: you would need to copy the scene and
adjust every level's Explosion Scene property.

To help build interest in getting to know Godot in an introductory
workshop, we want to make it possible to override a world's Goober or
Player textures by just placing image files at a particular path in the
world's directory, without needing to adjust anything in the editor
itself.

Use ResourceFinder.load_world_assets() to look for Player, Goober and
Explosion textures in a directory adjacent to the current scene. If any
are found, pick one at random for each instance. If none are found,
preserve the old behaviour.

Document this improvement in the WorldBuilding.md file.