CandySpawner: Cycle through all textures
Previously, sprites were created only when a new candy needs to fall and
the idle list is empty, with unused sprites placed there once they fall
off the bottom of the window.

This was fine with only one texture; but now, multiple textures are
supported. Although those textures are shuffled, on the first level the
interval between drops means only one candy is on-screen at a time, so
only one texture is used. If you have many textures (perhaps from
different learners) you might never see more than one or two of them
without completing the game.

Instead, prefill the idle list with as many sprites as there are
textures; and treat it as FIFO rather than LIFO.