Promote breakable brick from Cgm-Dali
This is a nice component for worlds beyond Cgm-Dali to use.

Move it to the top-level Scene and Script directories, renaming to
PascalCase to match the rest of the project (which is not the standard
Godot style but there we go). Add the scene to the default TileSet so it
can easily be used in other worlds.