CandyWrapper: Add placeholder image directories
This allows quick replacement of the assets in the base game.

We don't intend that any replacement assets in these directories will
actually be merged to main; the intended use is in short workshops where
learners are introduced to how to modify the game. Unlike replacing the
images in the sample world, the original game has enough levels – and
difficulty – to be played for more than a few minutes, which will allow
more assets to be randomly selected during play.