ci: Remove check for whether GitHub Pages is enabled
I put this check into the ancestor of this workflow partly to provide a
hint when someone inexperienced with GitHub Actions & Pages tries to
reuse the workflow, but mainly to avoid wasting resources building the
web version of the game if it's not going to be used.

However, the web version of the game is now built unconditionally in the
'export.yml' workflow. And in practice this check always succeeds; if it
fails, the developer setting up the workflow will figure it out. So it
is currently just adding ~10 seconds of extra time to the build process
for no benefit.

Remove it.