Relabel “Extra Worlds” as “Community Worlds”
This more clearly identifies the worlds as being a community-driven project.

Leave the “Extra” terminology in filenames and code for the menu:
renaming them is churn, and I don't expect this to cause any confusion
for contributors, who don't need to touch the menu code.

See https://github.com/endlessm/everlasting-candy/issues/126