World One: Desaturate all assets
With this change, all colour is removed from World One. The idea is
that, as learners add their own artwork, colour will return to the game.

To desaturate the assets themselves, I used the GNU Image Manipulation
Program's Colours → Desaturate → Desaturate tool.

For the Candy, Explosion, Goober and Player, I used the feature added in
https://github.com/endlessm/candy-collective/pull/95 that picks up
assets from the world directory. This has the disadvantage that, when
learners' colourful assets are dropped into World One, the desaturated
assets will still be included in the mix.

The alternative would be to add a copy of the Player, Goober and
Explosion scene in the World One directory and adjust each scene to use
the desaturated assets. As well as cluttering the directory, this
actually would have a worse consequence: when a learner creates their
own world by copying the World One directory to Worlds/Unicorn Zone, the
copy of these scenes would still reference the *old* asset path:

    World One                       Unicorn Zone
    ---------                       ------------

    Player.tscn            -------- Player.tscn
        ↓                 /
    Player.png  <---------          Player.png

So the learner would find themselves editing their copy of the Player
asset and it having no effect in their world.

For essentially the same reason, the desaturated versions of the tiles
and the title texts are stored outside the World One directory. Otherwise,
again, if TileSet.tres and Tilemap.png were stored in the World One directory,
when the directory is copied, the new directory's TileSet.tres would
still reference the old path to the PNG, and the levels would reference
the World One TileSet.tres:

    World One                       Unicorn Zone
    ---------                       ------------

    1.tscn              ----------- 1.tscn
        ↓              /
    TileSet.tres <-----    -------- TileSet.tres
        ↓                 /
    Tilemap.png  <---------         Tilemap.png

I think it is less confusing that the TileSet.tres and Tilemap.png are
stored outside the World One directory. When a learner wants to modify
the tiles, they'll still have to update each level to refer to their new
tileset, but at least there is not a misleading unused copy in their
world already.

The desaturated version of Title.png has also been edited to refer to
World One rather than Candy Wrapper. There is no font for the
hand-lettered text but fortunately all of the letters of WORLD ONE are
present in the existing asset so I laboriously copied them around in the
GIMP!

Fixes https://github.com/endlessm/candy-collective/issues/128