ResourceFinder: Use ResourceLoader.list_directory() to find world assets
This new function, added in Godot 4.4, avoids the issue that resources
(like images) must be loaded by their original path, while in the
exported game DirAccess yields the real filenames used when the resource
has been converted during export.

Replace the use of DirAccess in the helper function with
ResourceLoader.list_directory(). While it's not documented,
experimentally any directories in the given path are indicated by a
trailing slash in the returned filename; use this in place of
.current_is_dir() to implement recursive search.